balance cypress split runs
