+++ web/template/voucher_start.gohtml
@@ -7,7 +7,7 @@
 
       {{ trHtml .App "voucherStartContent" }}
 
-      <p class="govuk-body test">
+      <p class="govuk-body">
         <a href="{{ link .App global.Paths.Voucher.Login.Format }}" role="button" draggable="false" class="govuk-button govuk-button--start" data-module="govuk-button">
           {{ tr .App "start" }}
           <svg class="govuk-button__start-icon" xmlns="http://www.w3.org/2000/svg" width="17.5" height="19" viewBox="0 0 33 40" aria-hidden="true" focusable="false">
